requirements.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0

# HTTP客户端和服务器
httpx==0.25.2
websockets==12.0

# 数据验证和处理
pydantic==2.5.0
python-multipart==0.0.6

# 音频处理
pyaudio==0.2.11
Pillow==10.1.0

# 阿里云百炼SDK
#dashscope==1.14.0

# Piper TTS
piper==1.2.0

# 日志和配置
python-dotenv==1.0.0
loguru==0.7.2

# 开发和测试
pytest==7.4.3
pytest-asyncio==0.21.1